[refactor/#22] 주석 추가 및 Dto 폴더링, Signup 함수명 수정 등
