+++ app/build.gradle
@@ -49,21 +49,21 @@ dependencies {
     implementation 'com.google.android.material:material:1.11.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.activity:activity:1.8.0'
-
+    // 서버통신을 위한 retrofit + retrofit에서 serialization 사용
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
-    implementation 'org.jetbrains.kotlinx:kotlinx-serialization-json:1.4.1'
     implementation 'com.jakewharton.retrofit:retrofit2-kotlinx-serialization-converter:1.0.0'
-
-    // define a BOM and its version
+    // json 직렬화
+    implementation 'org.jetbrains.kotlinx:kotlinx-serialization-json:1.4.1'
+    // OkHttp 버전 관리및 로깅 인터셉터
     implementation(platform("com.squareup.okhttp3:okhttp-bom:4.10.0"))
-
-    // define any required OkHttp artifacts without version
     implementation("com.squareup.okhttp3:okhttp")
     implementation("com.squareup.okhttp3:logging-interceptor")
+    // navigation 구성 요소에서 fragment / ui 사용
     implementation("androidx.navigation:navigation-fragment-ktx:2.7.7")
     implementation("androidx.navigation:navigation-ui-ktx:2.7.7")
+    // 단위 테스트
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    // ui 테스트
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
-
 }
