+++ app/src/main/java/com/sopt/now/presentation/Dto/RequestLoginDto.kt
@@ -1,0 +1,12 @@
+package com.sopt.now.presentation.Dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RequestLoginDto(
+    @SerialName("authenticationId")
+    val authenticationId: String,
+    @SerialName("password")
+    val password: String
+)
