+++ app/src/main/java/com/sopt/now/presentation/Dto/RequestSignUpDto.kt
@@ -1,12 +1,16 @@
-package com.sopt.now.presentation
+package com.sopt.now.presentation.Dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestLoginDto(
+data class RequestSignUpDto(
     @SerialName("authenticationId")
     val authenticationId: String,
     @SerialName("password")
-    val password: String
+    val password: String,
+    @SerialName("nickname")
+    val nickname: String,
+    @SerialName("phone")
+    val phone: String,
 )
