+++ app/src/main/java/com/sopt/now/presentation/Dto/ResponseLoginDto.kt
@@ -1,10 +1,10 @@
-package com.sopt.now.presentation
+package com.sopt.now.presentation.Dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponseSignUpDto(
+data class ResponseLoginDto(
     @SerialName("code")
     val code: Int,
     @SerialName("message")
