+++ app/src/main/java/com/sopt/now/presentation/Dto/ResponseSignUpDto.kt
@@ -1,10 +1,10 @@
-package com.sopt.now.presentation
+package com.sopt.now.presentation.Dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponseLoginDto(
+data class ResponseSignUpDto(
     @SerialName("code")
     val code: Int,
     @SerialName("message")
