+++ app/src/main/java/com/sopt/now/presentation/Dto/ResponseUserInfoDto.kt
@@ -1,14 +1,22 @@
-package com.sopt.now.presentation
+package com.sopt.now.presentation.Dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestSignUpDto(
+data class ResponseUserInfoDto(
+    @SerialName("code")
+    val code: Int,
+    @SerialName("message")
+    val message: String,
+    @SerialName("data")
+    val data: UserData,
+)
+
+@Serializable
+data class UserData(
     @SerialName("authenticationId")
     val authenticationId: String,
-    @SerialName("password")
-    val password: String,
     @SerialName("nickname")
     val nickname: String,
     @SerialName("phone")
