--- app/src/main/java/com/sopt/now/presentation/ResponseUserInfoDto.kt
@@ -1,24 +1,0 @@
-package com.sopt.now.presentation
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResponseUserInfoDto(
-    @SerialName("code")
-    val code: Int,
-    @SerialName("message")
-    val message: String,
-    @SerialName("data")
-    val data: UserData,
-)
-
-@Serializable
-data class UserData(
-    @SerialName("authenticationId")
-    val authenticationId: String,
-    @SerialName("nickname")
-    val nickname: String,
-    @SerialName("phone")
-    val phone: String,
-)
